local changes

+++ front/app/components/SiteItem/SiteViewItem.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { Button, Checkbox } from 'react-bootstrap';
+import { Button, Checkbox, FormControl } from 'react-bootstrap';
 import { SiteViewFragment } from 'types/SiteViewFragment';
 import { withRouter } from 'react-router-dom';
 import DeleteSiteViewMutation, {
@@ -11,6 +11,7 @@ import UpdateSiteViewMutation, {
   UpdateSiteViewMutationFn,
 } from 'mutations/UpdateSiteViewMutation';
 import { view, update } from 'ramda';
+import { updateView } from 'utils/siteViewUpdater';
 
 interface SiteViewItemProps {
   match: any;
@@ -20,11 +21,29 @@ interface SiteViewItemProps {
   siteView: SiteViewFragment;
 }
 
+interface SiteViewItemState {
+  form: {
+    siteViewName: string;
+    siteViewPath: string;
+  };
+  rename: boolean;
+}
+
 const StyledButton = styled(Button)`
   margin-right: 15px;
 `;
 
-class SiteViewItem extends React.PureComponent<SiteViewItemProps> {
+class SiteViewItem extends React.PureComponent<
+  SiteViewItemProps,
+  SiteViewItemState
+> {
+  state: SiteViewItemState = {
+    form: {
+      siteViewName: '',
+      siteViewPath: '',
+    },
+    rename: false,
+  };
   handleEditClick = () => {
     const siteViewId = this.props.siteView.id;
     const siteId = this.props.match.params.id;
@@ -33,9 +52,35 @@ class SiteViewItem extends React.PureComponent<SiteViewItemProps> {
     );
   };
 
+  handleRenameClick = (updateSiteView: UpdateSiteViewMutationFn) => {
+    const { siteView } = this.props;
+    const { form, rename } = this.state;
+
+    this.setState(
+      {
+        rename: true,
+      },
+      () => {
+        updateSiteView({
+          variables: {
+            input: {
+              id: siteView.id,
+              name: form.siteViewName,
+              url: form.siteViewPath,
+              mutations: [],
+            },
+          },
+        }).then(this.props.refresh());
+      }
+    );
+    this.
+    )
+  };
+
   handleCheckbox = (updateSiteView: UpdateSiteViewMutationFn) => {
     const { siteView } = this.props;
     if (siteView.default) {
+      alert('There must be a default site view.');
       return null;
     }
     updateSiteView({
@@ -53,12 +98,17 @@ class SiteViewItem extends React.PureComponent<SiteViewItemProps> {
   };
 
   handleDelete = (deleteSiteView: DeleteSiteViewMutationFn) => {
+    const { siteView } = this.props;
+    if (siteView.default) {
+      alert('There must be a default site.');
+      return null;
+    }
     if (!window) return;
     if (window.confirm('Are you sure?')) {
       deleteSiteView({
         variables: {
           input: {
-            id: this.props.siteView.id,
+            id: siteView.id,
           },
         },
       }).then(res => {
@@ -67,12 +117,43 @@ class SiteViewItem extends React.PureComponent<SiteViewItemProps> {
       });
     }
   };
+
+  handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    this.setState({
+      form: { ...this.state.form, [e.target.name as any]: e.target.value },
+    });
+  };
+
   render() {
     const { siteView } = this.props;
+    const { rename } = this.state;
     return (
-      <tr>
-        <td>{siteView.name}</td>
-        <td>{siteView.url}</td>
+      <>
+        {rename ? (
+          <tr>
+            <td>
+              <FormControl
+                name="siteViewName"
+                placeholder="Site Name"
+                value={siteView.name}
+                onChange={this.handleInputChange}
+              />
+            </td>
+            <td>
+              <FormControl
+                name="siteViewPath"
+                placeholder="Site View Path"
+                value={siteView.url}
+                onChange={this.handleInputChange}
+              />
+            </td>
+          </tr>
+        ) : (
+          <tr>
+            <td>{siteView.name}</td>
+            <td>{siteView.url}</td>
+          </tr>
+        )}
         <td>
           <UpdateSiteViewMutation>
             {updateSiteView => (
@@ -92,8 +173,9 @@ class SiteViewItem extends React.PureComponent<SiteViewItemProps> {
               </StyledButton>
             )}
           </DeleteSiteViewMutation>
+          <StyledButton onClick={this.handleRenameClick}>Rename</StyledButton>
         </td>
-      </tr>
+      </>
     );
   }
 }
